+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.1.0 - 2026-02-10]
+## [1.1.0 - 2026-02-12]
 ### Added
 - Dockerfile
 - Single Spa
@@ -110,6 +110,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Download Form Template PDF Buttons
 - Selected Forms as Project View Row
 - Frontend Name
+- Project and Forms Configuration
 
 ### Changed
 - Rename accept and reject bridgehead state buttons (authorize/revoke)